iss #59: (🎓 go) wip: parallel parser: integrate parallel scan for files
