+++ fb2_scanner/fs/fs.go
@@ -3,8 +3,12 @@ package fs
 import (
 	"errors"
 	"fmt"
+	"os"
+	"sync"
+	"time"
 
 	"dev.maizy.ru/ponylib/fb2_parser"
+	"dev.maizy.ru/ponylib/fb2_parser/metadata"
 	"dev.maizy.ru/ponylib/fb2_scanner"
 	"dev.maizy.ru/ponylib/internal/u"
 )
@@ -29,14 +33,12 @@ func (f *FileTarget) Type() fb2_scanner.TargetType {
 	return fb2_scanner.FsFile
 }
 
-func (f *FileTarget) Scan() <-chan fb2_scanner.ScannerResult {
+func (f *FileTarget) Scan(openFileSemaphore chan struct{}) <-chan fb2_scanner.ScannerResult {
 	resultChan := make(chan fb2_scanner.ScannerResult)
+	wg := sync.WaitGroup{}
+	scanRegularFile(f.Path, resultChan, &wg, openFileSemaphore)
 	go func() {
-		resultChan <- fb2_scanner.ScannerResult{
-			Source:   &FileSource{"../singele-fake-file.fb2"},
-			Metadata: &fb2_parser.Fb2Metadata{Book: &fb2_parser.Book{Title: u.StrPtr("title single")}},
-			Error:    nil,
-		}
+		wg.Wait()
 		close(resultChan)
 	}()
 	return resultChan
@@ -54,7 +56,7 @@ func (f *DirectoryTarget) Type() fb2_scanner.TargetType {
 	return fb2_scanner.FsDir
 }
 
-func (f *DirectoryTarget) Scan() <-chan fb2_scanner.ScannerResult {
+func (f *DirectoryTarget) Scan(openFileSemaphore chan struct{}) <-chan fb2_scanner.ScannerResult {
 	resultChan := make(chan fb2_scanner.ScannerResult)
 	go func() {
 		defer close(resultChan)
@@ -81,4 +83,69 @@ func (f *DirectoryTarget) Scan() <-chan fb2_scanner.ScannerResult {
 		}
 	}()
 	return resultChan
+}
+
+func scanRegularFile(
+	path string, resultChan chan<- fb2_scanner.ScannerResult, wg *sync.WaitGroup, openFileSemaphore chan struct{}) {
+
+	source := &FileSource{path}
+	stat, err := os.Stat(path)
+	wg.Add(1)
+	if err != nil {
+		resultChan <- fb2_scanner.ScannerResult{
+			Source:   source,
+			Metadata: nil,
+			Error:    u.ErrPtr(fmt.Errorf("unable to open %s: %w", path, err)),
+		}
+		wg.Done()
+		return
+	}
+	if !stat.Mode().IsRegular() {
+		resultChan <- fb2_scanner.ScannerResult{
+			Source:   source,
+			Metadata: nil,
+			Error:    u.ErrPtr(fmt.Errorf("%s isn't a regular file", path)),
+		}
+		wg.Done()
+		return
+	}
+	openFileSemaphore <- struct{}{}
+	go func() {
+		fp, err := os.Open(path)
+		defer func() {
+			_ = fp.Close()
+		}()
+		defer wg.Done()
+
+		if err != nil {
+			resultChan <- fb2_scanner.ScannerResult{
+				Source:   source,
+				Metadata: nil,
+				Error:    u.ErrPtr(fmt.Errorf("unable to open %s: %w", path, err)),
+			}
+			return
+		}
+
+		parseStart := time.Now()
+		parseMetadata, err := metadata.ParseMetadata(fp)
+		timers := fb2_scanner.ParseTimers{ParseTimeNs: time.Since(parseStart).Nanoseconds()}
+
+		if err != nil {
+			resultChan <- fb2_scanner.ScannerResult{
+				Source:   source,
+				Metadata: nil,
+				Error:    u.ErrPtr(fmt.Errorf("unable to open %s: %w", path, err)),
+				Timers:   timers,
+			}
+			return
+		}
+
+		resultChan <- fb2_scanner.ScannerResult{
+			Source:   source,
+			Metadata: parseMetadata,
+			Error:    nil,
+			Timers:   timers,
+		}
+		<-openFileSemaphore
+	}()
 }
