+++ internal/u/ptr.go
@@ -7,3 +7,7 @@ func StrPtr(value string) *string {
 func IntPtr(value int) *int {
 	return &value
 }
+
+func ErrPtr(error error) *error {
+	return &error
+}
